% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_reliabilities}
\alias{add_reliabilities}
\title{Add item reliabilities to a multiverse pipeline}
\usage{
add_reliabilities(.df, scale_name, items)
}
\arguments{
\item{.df}{the original \code{data.frame}(e.g., base data set). If part of
set of add_* decision functions in a pipeline, the base data will be passed
along as an attribute.}

\item{scale_name}{a character string. Indicates the name of the scale or
measure measured by the items or indicators in \code{items}.}

\item{items}{the items (variables) that comprise a scale or measure. These
variables will be passed to \code{link[performance]{cronbachs_alpha}},
\code{link[performance]{item_intercor}}, and
\code{link[performance]{item_reliability}}. You can also use tidyselect to
select variables.}
}
\value{
a \code{data.frame}with three columns: type, group, and code. Type
indicates the decision type, group is a decision, and the code is the
actual code that will be executed. If part of a pipe, the current set of
decisions will be appended as new rows.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This function has been
deprecated; please use \code{\link[=add_model_descriptives]{add_model_descriptives()}} instead.
}
\examples{

library(tidyverse)
library(multitool)

the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

the_data |>
  add_filters(include1 == 0,include2 != 3,include2 != 2, include3 > -2.5) |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_reliabilities("unp_scale", c(iv1,iv2,iv3))
}
