% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Create a blueprint figure of your multiverse pipeline}
\usage{
blueprint(.grid, splines = "line", graph = TRUE, show_code = FALSE, ...)
}
\arguments{
\item{.grid}{a pipeline \code{tibble}}

\item{splines}{how to draw edges connecting nodes in Graphviz, defaults to
\code{splines = "line"}}

\item{graph}{logical, whether or not to draw the graph in the viewing pane}

\item{show_code}{logical, whether to print out the dot code to the console}

\item{...}{additional arguments passed to \code{\link[DiagrammeR]{grViz}}}
}
\value{
if \code{graph = TRUE}, the object will be a Graphviz diagram
generated by \code{\link[DiagrammeR]{grViz}}. If \code{graph = FALSE}, the output
will be the literal dot code to generate the graph, which could be used to
manually edit or customize the graph further.
}
\description{
Create a blueprint figure of your multiverse pipeline
}
\examples{

library(tidyverse)
library(multitool)

## Simulate some data
the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

full_pipeline <-
  the_data |>
  add_filters(include1 == 0,include2 != 3,include2 != 2,scale(include3) > -2.5) |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_preprocess(process_name = "Standardize IVs", 'mutate({ivs} = scale({ivs}))') |>
  add_preprocess(process_name = "Standardize Moderators", mutate({mods} := scale({mods}))) |>
  add_summary_stats("IV Summary Stats", starts_with("iv"), c("mean", "sd")) |>
  add_summary_stats("DV Summary Stats", starts_with("dv"), c("skewness", "kurtosis")) |>
  add_correlations("Between Predictors", matches("iv|mod|cov"), focus_set = c(cov1,cov2)) |>
  add_correlations("Between Outcomes", matches("dv|mod"), focus_set = matches("dv")) |>
  add_cron_alpha("unp_scale", c(iv1,iv2,iv3)) |>
  add_cron_alpha("vio_scale", starts_with("mod")) |>
  add_model("no covariates",lm({dvs} ~ {ivs} * {mods})) |>
  add_model("covariates",lm({dvs} ~ {ivs} * {mods} + cov1)) |>
  add_postprocess("simple slopes", "sim_slopes(pred = condition, modx = {ivs}, modx_values = c(-1,1))") |>
  add_postprocess("plotting interactions", ggpredict(terms = c("{ivs} [-1,1]", "{mods} [-1,1]")))

# Default graph produced
blueprint(full_pipeline)

# Change the way edges look
blueprint(full_pipeline, splines = "polyline")

# Take a look at the dot code
blueprint(full_pipeline, graph = FALSE, show_code = TRUE)
}
