% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{create_corr_grid}
\alias{create_corr_grid}
\title{Title}
\usage{
create_corr_grid(.df, corr_vars, focus_vars, add_stretch = NULL)
}
\arguments{
\item{.df}{blah blah}

\item{corr_vars}{blah blah}

\item{focus_vars}{blah blah}

\item{add_stretch}{blah blah}
}
\value{
a sum
}
\description{
Title
}
\examples{
sum(c(1,2,3,4))
}
