% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{create_desc_grid}
\alias{create_desc_grid}
\title{Title}
\usage{
create_desc_grid(.df, ..., stats)
}
\arguments{
\item{.df}{the base dataset}

\item{...}{blah blah}

\item{focus_vars}{blah blah}

\item{add_stretch}{blah blah}
}
\value{
a correlation grid
}
\description{
Title
}
\examples{
sum(c(1,2,3,4))
}
