% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{create_filter_grid}
\alias{create_filter_grid}
\title{Create all combinations of filtering decisions for exploring exclusions}
\usage{
create_filter_grid(my_data, ..., print = TRUE)
}
\arguments{
\item{my_data}{the actual data as a \code{data.frame} you want to create a
filtering decisions for}

\item{...}{logical expressions to be used with \code{\link[dplyr]{filter}}
separated by commas. Expressions should not be quoted.}

\item{print}{logical, whether or not to show a summary of the resulting grid.
Defaults to \code{TRUE}}
}
\value{
a list with two components. The first is a summary of your filters
with the following columns: \code{expr}, \code{expr_var}, \code{expr_n},
and \code{expr_typ}e:

\describe{ \item{\code{expr}}{the expression(s) passed to
\code{create_filter_grid()}} \item{\code{expr_var}}{the variable involved
in the filtering expression} \item{\code{expr_n}}{the total number of rows
after the expression is applied} \item{\code{expr_type}}{whether or not the
expression filters data} }

The summary is meant to help the user understand and evaluate different
filters. The function automatically generates expressions that do nothing
to complete a set of alternatives.

The second component of the list is a full grid of filtering decisions.
This is created under the hood by \code{\link[tidyr]{expand_grid}}.
}
\description{
Create all combinations of filtering decisions for exploring exclusions
}
\examples{
library(multitool)

my_data <-
  data.frame(
   id   = 1:500,
   iv1  = rnorm(500),
   iv2  = rnorm(500),
   iv3  = rnorm(500),
   covariate1 = rnorm(500),
   covariate2 = rnorm(500),
   dv1 = rnorm(500),
   dv2 = rnorm(500),
   filter1   = sample(1:3, size = 500, replace = TRUE),
   filter2   = rnorm(500),
   filter3   = rbinom(500, size = 1, prob = .1),
   filter4   = rbinom(500, size = 1, prob = .1)
  )

my_filter_grid <-
  create_filter_grid(
    my_data,
    filter1 == 1,
    filter1 == 2,
    filter2 == 0,
    filter3 == 0,
    filter4 == 0
  )
}
