% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{create_model_grid}
\alias{create_model_grid}
\title{Create a modeling grid}
\usage{
create_model_grid(...)
}
\arguments{
\item{...}{literal model syntax (no quotes) you would like to run. You can
use \code{glue} inside formulas to dynamically generate variable names
based on a variable grid. For example, if you make variable grid with two
versions of your IVs (e.g., \code{iv1} and \code{iv2}), you can write your
formula like so: \code{lm(happiness ~ {iv} + control_var)}. The only
requirement is that the variables written in the formula actually exist in
the underlying data. You are also responsible for loading any packages that
run a particular model (e.g., \code{lme4} for mixed-models)}
}
\value{
\ a list with four components:

\describe{
\item{summary}{a summary \code{tibble} of the models chosen.}
\item{combinations}{a summary of the number of alternatives models}
\item{n_combinations}{the total number of model combinations}
\item{grid}{the actual modeling grid, the number of rows = n_combinations.
}
}
}
\description{
Create a modeling grid
}
\examples{
library(tidyverse)
library(multitool)

my_model_grid <-
  create_model_grid(
    lm({dv} ~ {iv} * {mod}),
    lm({dv} ~ {iv} * {mod} + {cov})
  )
}
