% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icar_data_codebook.R
\docType{data}
\name{icar_data_codebook}
\alias{icar_data_codebook}
\title{A variable, variable label, and value label codebook for the ICAR dataset}
\format{
A data frame with {63} rows and {3} variables:
\describe{
\item{var_name}{name of the variable in the data}
\item{var_label}{brief description of each variable}
\item{var_values}{explanation of any values and labels where applicable}
}
}
\source{
\url{data_URL}
}
\usage{
icar_data_codebook
}
\description{
This data frame contains the same information as the documentation for the
ICAR dataset but provides it for use in the session.
}
\keyword{datasets}
