% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{model_summary_code}
\alias{model_summary_code}
\title{Add a model summary function to the grid}
\usage{
model_summary_code(...)
}
\arguments{
\item{...}{the literal code (unquoted) you would like to execute on the model
object. This usually \code{\link[broom]{tidy}} but could be other
functions.

The code should be written to work with pipes (i.e., \code{|>} or
\code{\%>\%}) because the model object will be passed directly to the
summary function of choice.

So if you fit a simple linear model like: \code{lm(y ~ x1 + x2)}, and you
want to save a summary instead of the entire \code{lm} object, would simply
pass \code{summary()} to \code{model_summary_code()}. The underlying code
would be:

\code{data |> filters |> lm(y ~ x1 + x2, data = _) |> summary()}}
}
\value{
a \code{tibble} with  two columns, "summary_code" and "code".
}
\description{
Add a model summary function to the grid
}
\examples{

library(tidyverse)
library(multitool)

my_model_summary_code <- model_summary_code(tidy())
}
