% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{post_filter_code}
\alias{post_filter_code}
\title{Add arbitrary code to execute after data are filtered (but before analysis)}
\usage{
post_filter_code(grid, ...)
}
\arguments{
\item{grid}{a \code{tibble} produced by \code{\link{combine_all_grids}}}

\item{...}{the literal code you would like to execute after data are
filtered. \code{glue} syntax is allowed. An example might be centering or
scaling a predictor after the appropriate filters are applied to the data.}
}
\value{
Your combined grid with a new column containing arbitrary code to run
post filtering but pre-analysis
}
\description{
Add arbitrary code to execute after data are filtered (but before analysis)
}
\examples{
library(multitool)

my_data <-
  data.frame(
   id   = 1:500,
   iv1  = rnorm(500),
   iv2  = rnorm(500),
   iv3  = rnorm(500),
   covariate1 = rnorm(500),
   covariate2 = rnorm(500),
   dv1 = rnorm(500),
   dv2 = rnorm(500),
   filter1   = sample(1:3, size = 500, replace = TRUE),
   filter2   = rnorm(500),
   filter3   = rbinom(500, size = 1, prob = .1),
   filter4   = rbinom(500, size = 1, prob = .1)
  )

my_filter_grid <-
  create_filter_grid(
    my_data,
    filter1 == 1,
    filter1 == 2,
    filter2 == 0,
    filter3 == 0,
    filter4 == 0
  )



post_filter_code(my_filter_grid, mutate({iv} = scale({iv}) |> as.numeric()))

}
