% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model-processing.R
\name{postprocess_model}
\alias{postprocess_model}
\title{Post-process a fitted model}
\usage{
postprocess_model(model, code, additional_args = "")
}
\description{
Post-process a fitted model
}
\keyword{internal}
