% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_analysis.R
\name{preanalyze_df}
\alias{preanalyze_df}
\title{Preanalysis of Data Frame}
\usage{
preanalyze_df(df, index.limit = 3)
}
\arguments{
\item{df}{A Data Frame.}

\item{index.limit}{How big combinations of variables to consider as index.}
}
\value{
An object off class "MltDataFrame".
}
\description{
Returns a MltDataFrame object obtained from given Data Frame, containing possible indexes as variable combinations and types predictions.
}
