% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{reveal_model_warnings}
\alias{reveal_model_warnings}
\title{Reveal any warnings about your models during a multiverse analysis}
\usage{
reveal_model_warnings(.multi, .unpack_specs = "no")
}
\arguments{
\item{.multi}{a multiverse list-column \code{tibble} produced by
\code{\link{run_multiverse}}.}

\item{.unpack_specs}{character, options are \code{"no"}, \code{"wide"}, or
\code{"long"}. \code{"no"} (default) keeps specifications in a list column,
\code{wide} unnests specifications with each specification category as a
column. \code{"long"} unnests specifications and stacks them into long
format, which stacks specifications into a \code{decision_set} and
\code{alternatives} columns. This is mainly useful for plotting.}
}
\value{
the unnested model warnings captured during analysis
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This function has been
deprecated; please use \code{\link[=unpack_model_warnings]{unpack_model_warnings()}} instead.
}
\examples{

library(tidyverse)
library(multitool)

# Simulate some data
the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

# Decision pipeline
full_pipeline <-
  the_data |>
  add_filters(include1 == 0,include2 != 3,include2 != 2,scale(include3) > -2.5) |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_model("linear_model", lm({dvs} ~ {ivs} * {mods} + cov1))

pipeline_grid <- expand_decisions(full_pipeline)

# Run the whole multiverse
the_multiverse <- run_multiverse(pipeline_grid[1:10,])

# Reveal results of the linear model
the_multiverse |>
  reveal_model_warnings()
}
