% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-multiverse.R
\name{run_universe}
\alias{run_universe}
\title{Run a single set of arbitrary decisions and save the result}
\usage{
run_universe(.grid, decision_num, save_model = FALSE)
}
\arguments{
\item{.grid}{a \code{tibble} produced by \code{\link{combine_all_grids}}}

\item{decision_num}{an single integer from 1 to \code{nrow(grid)} indicating
which specific decision set to run}

\item{save_model}{logical, indicates whether to save the model object in its
entirety. The default is \code{FALSE} because model objects are usually
large and under the hood, \code{\link[broom]{tidy}} and
\code{\link[broom]{glance}} is used to summarize the most useful model
information.}
}
\value{
a single row \code{tibble} containing the decision, code that was
ran, the results, and any notes (e.g., warnings or messages).
}
\description{
Run a single set of arbitrary decisions and save the result
}
\examples{

\dontrun{
run_universe(.grid, .df, decision_num)
}
}
