% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-pipelines.R
\name{show_code}
\alias{show_code}
\alias{show_code_subgroups}
\alias{show_code_filters}
\alias{show_code_preprocess}
\alias{show_code_model}
\alias{show_code_postprocess}
\alias{show_code_summary_stats}
\alias{show_code_corrs}
\alias{show_code_reliabilities}
\title{Show multiverse data code pipelines}
\usage{
show_code(
  .grid,
  decision_num,
  .step = "model",
  .post_step = NULL,
  .execute = FALSE
)

show_code_subgroups(.grid, decision_num, ...)

show_code_filters(.grid, decision_num, ...)

show_code_preprocess(.grid, decision_num, ...)

show_code_model(.grid, decision_num, ...)

show_code_postprocess(.grid, decision_num, ...)

show_code_summary_stats(
  .grid,
  decision_num,
  summary_set = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_corrs(
  .grid,
  decision_num,
  corr_set = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)

show_code_reliabilities(
  .grid,
  decision_num,
  rel_set = 1,
  copy = FALSE,
  console = TRUE,
  execute = FALSE,
  ...
)
}
\arguments{
\item{.grid}{a full decision grid created by \code{\link{expand_decisions}}
or a fully analyzed grid produced by \code{\link{analyze_grid}}.}

\item{decision_num}{numeric. Indicates which decision set in the grid to show
underlying code.}

\item{.step}{a point along the pipeline for which you would like to show the
underlying code. Defaults to the model.}

\item{.post_step}{Only relevant if you are exposing a postprocessing step. If
you have more than one postprocess, you can specify which you would like to
expose by index or by name.}

\item{.execute}{logical, whether or not to run the code as well as print it.}

\item{...}{additional arguments passed to \code{show_code()}}

\item{summary_set}{numeric. For \code{show_code_summary_stats}, Which set of
summary statistics to print. Default is set to the \code{1}.}

\item{copy}{logical, whether to copy code to clipboard}

\item{console}{logical, whether to paste code into the console}

\item{execute}{logical, whether to run the code}

\item{corr_set}{numeric. For \code{show_code_corrs}, Which set of
correlations to print. Default is set to the \code{1}.}

\item{rel_set}{numeric. For \code{show_code_reliabilities}, Which set of
reliabilities to print. Default is set to the \code{1}.}
}
\value{
the code that generated results up to the specified point in an
analysis pipeline.
}
\description{
\code{show_code} is the generic function. All \code{show_code*} functions are
simple wrappers of \code{show_code}.
}
\details{
Each \code{show_code*} function should be self-explanatory - they indicate
where along the multiverse pipeline to extract code. The goal of these
functions is to create a window into each data/model combination and allow
the user to inspect specific decisions straight from the code that produced
it.
}
\section{Functions}{
\itemize{
\item \code{show_code_subgroups()}: Show the code up to the subgroups stage

\item \code{show_code_filters()}: Show the code up to the filtering stage

\item \code{show_code_preprocess()}: Show the code up to the preprocessing stage

\item \code{show_code_model()}: Show the code up to the modeling stage

\item \code{show_code_postprocess()}: Show the code up to the post-processing stage

\item \code{show_code_summary_stats()}: Show the code for computing summary statistics

\item \code{show_code_corrs()}: Show the code for computing correlations

\item \code{show_code_reliabilities()}: Show the code for computing scale reliability

}}
