% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-describe.R
\name{summarize_filter_ns}
\alias{summarize_filter_ns}
\title{Summarize samples sizes for each unique filtering expression}
\usage{
summarize_filter_ns(.pipeline)
}
\arguments{
\item{.pipeline}{a \code{data.frame} produced by calling a series of add_*
functions.}
}
\value{
a \code{tibble} with each row representing a filtering expression and
four columns: \code{filter_expression}, \code{variable}, \code{n_retained},
and \code{n_excluded}.
}
\description{
Summarize samples sizes for each unique filtering expression
}
\examples{
library(tidyverse)
library(multitool)

# create some data
the_data <-
  data.frame(
    id  = 1:500,
    iv1 = rnorm(500),
    iv2 = rnorm(500),
    iv3 = rnorm(500),
    mod = rnorm(500),
    dv1 = rnorm(500),
    dv2 = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

# create a pipeline blueprint
full_pipeline <-
  the_data |>
  add_filters(include1 == 0, include2 != 3, include3 > -2.5) |>
  add_variables(var_group = "ivs", iv1, iv2, iv3) |>
  add_variables(var_group = "dvs", dv1, dv2) |>
  add_model("linear model", lm({dvs} ~ {ivs} * mod))

summarize_filter_ns(full_pipeline)
}
