% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack-multiverse.R
\name{unpack_results}
\alias{unpack_results}
\alias{unpack_model_parameters}
\alias{unpack_model_performance}
\alias{unpack_model_warnings}
\alias{unpack_model_messges}
\alias{unpack_postprocess}
\title{Unpack a component of your analyzed grid}
\usage{
unpack_results(.multi, .what, .which = NULL, .unpack_specs = "wide")

unpack_model_parameters(.multi, effect_key = NULL, .unpack_specs = "wide")

unpack_model_performance(.multi, .unpack_specs = "wide")

unpack_model_warnings(.multi, .unpack_specs = "wide")

unpack_model_messges(.multi, .unpack_specs = "wide")

unpack_postprocess(.multi, .which, .unpack_specs = "wide")
}
\arguments{
\item{.multi}{a multiverse list-column \code{tibble} produced by
\code{\link{analyze_grid}}.}

\item{.what}{the name of a list-column you would like to unpack}

\item{.which}{any sub-list columns you would like to unpack}

\item{.unpack_specs}{character, options are \code{"no"}, \code{"wide"}, or
\code{"long"}. \code{"no"} (default) keeps specifications in a list column,
\code{wide} unnests specifications with each specification category as a
column. \code{"long"} unnests specifications and stacks them into long
format, which stacks specifications into a \code{decision_type},
\code{decision_set} and \code{decision_choice} columns. This is mainly
useful for plotting.}

\item{effect_key}{character, if you added parameter keys to your pipeline,
you can specify if you would like filter the parameters using one of your
parameter keys. This is useful when different variables are being switched
out across the multiverse but represent the same effect of interest.}
}
\value{
the unnested part of the multiverse requested. This usually contains
the particular estimates or statistics you would like to analyze over the
decision grid specified.
}
\description{
Unpack a component of your analyzed grid
}
\section{Functions}{
\itemize{
\item \code{unpack_model_parameters()}: Unpack the model parameters

\item \code{unpack_model_performance()}: Unpack the model performance

\item \code{unpack_model_warnings()}: Unpack the model warnings

\item \code{unpack_model_messges()}: Unpack the model messages

\item \code{unpack_postprocess()}: Unpack a post-processing result

}}
\examples{

library(tidyverse)
library(multitool)

# Simulate some data
the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

# Decision pipeline
full_pipeline <-
  the_data |>
  add_filters(include1 == 0,include2 != 3,include2 != 2,scale(include3) > -2.5) |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_model("linear_model", lm({dvs} ~ {ivs} * {mods} + cov1))

pipeline_grid <- expand_decisions(full_pipeline)

# Run the whole multiverse
the_multiverse <- analyze_grid(pipeline_grid[1:10,])

# Reveal results of the linear model
the_multiverse |> unpack_results(model_fitted, model_parameters)
}
