% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-pipeline.R
\name{add_summary_stats}
\alias{add_summary_stats}
\title{Add a set of descriptiove statistics to compute over a set of variables}
\usage{
add_summary_stats(.df, var_set, variables, stats)
}
\arguments{
\item{.df}{The original data.frame (e.g., base data set). If part of set of
add_* decision functions in a pipeline, the base data will be passed along
as an attribute.}

\item{var_set}{a character string. A name for the set of summary statistics}

\item{variables}{the variables for which you would like to compute summary
statistics. You can also use tidyselect to select variables.}

\item{stats}{a character vector of stat names (e.g., \code{c("mean","sd")}).
You are responsible for loading any packages that compute your preferred
summary statistics. Summary statistic functions must work inside
\code{\link[dplyr]{summarize}}.}
}
\value{
a data.frame with three columns: type, group, and code. Type
indicates the decision type, group is a decision, and the code is the
actual code that will be executed. If part of a pipe, the current set of
decisions will be appended as new rows.
}
\description{
Add a set of descriptiove statistics to compute over a set of variables
}
\examples{
library(tidyverse)
library(multitool)

the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

the_data |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_filters(include1 == 0, include2 != 3, scale(include3) > -2.5) |>
  add_summary_stats("iv_stats", starts_with("iv"), c("mean", "sd")) |>
  add_summary_stats("dv_stats", starts_with("dv"), c("skewness", "kurtosis"))
}
