% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-grids.R
\name{combine_all_grids}
\alias{combine_all_grids}
\title{Combine grids of analysis decisions into a single \code{tibble}}
\usage{
combine_all_grids(filter_grid = NULL, var_grid = NULL, model_grid = NULL)
}
\arguments{
\item{filter_grid}{a data.frame created using
\code{\link{create_filter_grid}}. This grid contains all observation
(row) filtering decisions.}

\item{var_grid}{a data.frame created using \code{\link{create_var_grid}}.
This grid should contain different versions of variables or sets of
different versions of variables.}

\item{model_grid}{a data.frame created using
\code{\link{create_model_grid}}. This grid contains the actual model
syntax (with possible alternatives) to run over the specified grids.}
}
\value{
The output of this function will be a \code{tibble} of all
combinations of analysis decisions. Each decision is denoted by the column
\code{decision}. Ecah type of decision grid will create a list column. For
example, when \code{filter_grid} is provided, the master grid will contain
the list column \code{filters} containing all filter deicions.
}
\description{
Combine grids of analysis decisions into a single \code{tibble}
}
\examples{
library(multitool)

my_data <-
  data.frame(
   id   = 1:500,
   iv1  = rnorm(500),
   iv2  = rnorm(500),
   iv3  = rnorm(500),
   covariate1 = rnorm(500),
   covariate2 = rnorm(500),
   dv1 = rnorm(500),
   dv2 = rnorm(500),
   filter1   = sample(1:3, size = 500, replace = TRUE),
   filter2   = rnorm(500),
   filter3   = rbinom(500, size = 1, prob = .1),
   filter4   = rbinom(500, size = 1, prob = .1)
  )

my_filter_grid <-
  create_filter_grid(
    my_data,
    filter1 == 1,
    filter1 == 2,
    filter2 == 0,
    filter3 == 0,
    filter4 == 0
  )

my_var_grid <-
  create_var_grid(
    my_data = my_data,
    iv       = c(iv1, iv2, iv3),
    dv       = c(dv1, dv2),
    covariates = c(covariate1, covariate2)
  )

my_model_grid <-
create_model_grid(
  lm({dv} ~ {iv}),
  lm({dv} ~ {iv} + {covariates})
)

combine_all_grids(my_filter_grid, my_var_grid, my_model_grid)

}
