% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-grids.R
\name{combine_all_grids}
\alias{combine_all_grids}
\title{Combine grids of analysis decisions into a single \code{tibble}}
\usage{
combine_all_grids(
  var_grid = NULL,
  filter_grid = NULL,
  model_grid = NULL,
  preprocessing = NULL,
  postprocessing = NULL
)
}
\arguments{
\item{var_grid}{a \code{tibble} created by \code{\link{create_var_grid}}.
This grid should contain different versions of variables or sets of
different versions of variables.}

\item{filter_grid}{a \code{tibble} created by
\code{\link{create_filter_grid}}. This grid contains all observation (row)
filtering decisions.}

\item{post_filter_code}{a \code{tibble} created by
\code{\link{post_filter_code}}. It should contain code you want to run just
after filtering but before analysis}

\item{model_code}{a \code{tibble} created by \code{\link{create_model_grid}}.
This grid contains the actual model syntax (with possible alternatives) to
run over the specified grids.}

\item{post_hoc_code}{a \code{tibble} created by \code{\link{post_hoc_code}}.
It should contain code that executes post hoc processing or analysis on the
model object specified in the \code{model_grid}. This could be something
like standardizing coefficients or running post hoc analysis (e.g., simple
slopes).}
}
\value{
The output of this function will be a \code{tibble} of all
combinations of analysis decisions. Each decision is denoted by the column
\code{decision}. Filter and variable grids (if they were included) will
become list columns. For example, when \code{filter_grid} is provided, the
master grid will contain the list column \code{filters} containing all
filter decisions. An analysis pipeline list column will also be created. In
the simplest case this will just be the model you plan to run. In more
complex cases, it will contain code to run after filters, the model,
different summaries, and any post-hoc tests.
}
\description{
Combine grids of analysis decisions into a single \code{tibble}
}
\examples{
library(multitool)
}
