% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack-multiverse.R
\name{condense}
\alias{condense}
\title{Summarize multiverse parameters}
\usage{
condense(.unpacked, .what, .how)
}
\arguments{
\item{.unpacked}{an unpacked (with \code{\link{reveal}} or
\code{\link{unnest}}) multiverse dataset.}

\item{.what}{a specific column to summarize. This could be a model estimate,
a summary statistic, correlation, or any other estimate computed over the
multiverse.}

\item{.how}{a named list. The list should contain summary functions (e.g.,
mean or median) the user would like to compute over the individual
estimates from the multiverse}
}
\value{
a summarized \code{tibble} containing a column for each summary
method from \code{.how}
}
\description{
Summarize multiverse parameters
}
\examples{
\dontrun{

library(tidyverse)
library(multitool)

# Simulate some data
the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

# Decision pipeline
full_pipeline <-
  the_data |>
  add_filters(include1 == 0,include2 != 3,include2 != 2,scale(include3) > -2.5) |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_preprocess(process_name = "scale_iv", 'mutate({ivs} = scale({ivs}))') |>
  add_preprocess(process_name = "scale_mod", mutate({mods} := scale({mods}))) |>
  add_summary_stats("iv_stats", starts_with("iv"), c("mean", "sd")) |>
  add_summary_stats("dv_stats", starts_with("dv"), c("skewness", "kurtosis")) |>
  add_correlations("predictors", matches("iv|mod|cov"), focus_set = c(cov1,cov2)) |>
  add_correlations("outcomes", matches("dv|mod"), focus_set = matches("dv")) |>
  add_cron_alpha("unp_scale", c(iv1,iv2,iv3)) |>
  add_cron_alpha("vio_scale", starts_with("mod")) |>
  add_model(lm({dvs} ~ {ivs} * {mods})) |>
  add_model(lm({dvs} ~ {ivs} * {mods} + cov1)) |>
  add_postprocess("aov", aov()) |>
  expand_decisions()

# Run the whole multiverse
the_multiverse <- run_multiverse(full_pipeline[1:10,])

# Reveal results of the linear model
the_multiverse |> reveal(lm_fitted, matches("tidy"), .unpack_specs = TRUE)

# Reveal and condense
the_multiverse |>
  reveal(lm_fitted, matches("tidy"), .unpack_specs = TRUE) |>
  group_by(term, dvs) |>
  condense(estimate, list(mn = mean, med = median))
  }
}
