% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{create_filter_grid}
\alias{create_filter_grid}
\title{Create all combinations of filtering decisions for exploring exclusion criteria}
\usage{
create_filter_grid(.df, ...)
}
\arguments{
\item{.df}{the actual data as a \code{data.frame} for which you want to create
filtering decisions}

\item{...}{logical expressions to be used with \code{\link[dplyr]{filter}}
separated by commas. Expressions should not be quoted.}
}
\value{
a list with four components:

\describe{
\item{summary}{a summary \code{tibble} of your filters with the following
columns: \code{filter_expr} (the filtering code), \code{filter_group}
(which variable does the filter apply to), \code{filtered_n}
(how many observations are there after the filter), and  \code{filter_type}
(does the expression filter or do nothing, which is auto-generated).}
\item{combinations}{a summary of the number of alternatives per filtering
decision}
\item{n_combinations}{the total number of filtering combinations}
\item{grid}{the actual filtering grid, the number of rows = n_combinations.
}
}
}
\description{
Create all combinations of filtering decisions for exploring exclusion criteria
}
\examples{
library(tidyverse)
library(multitool)

the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

my_filter_grid <-
  create_filter_grid(
    .df = the_data,
    include1 == 0,
    include2 != 3,
    include2 != 2,
    scale(include3) > -2.5
  )

my_filter_grid
}
