% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-pipeline.R
\name{expand_decisions}
\alias{expand_decisions}
\title{Expand a set of multiverse decisions into all possible combinations}
\usage{
expand_decisions(.pipeline, .pointer_path = NULL, .collect_after = NULL)
}
\arguments{
\item{.pipeline}{a \code{data.frame} produced by calling a series of add_*
functions.}

\item{.pointer_path}{a string specifying a path to create a external pointer
object. This is only necessary if you are using data from an external
source. Defaults to NULL.}

\item{.collect_after}{default is NULL. Most of the time you will not use this
argument. However, if your data come from a database, you can use this
argument to call \code{dplyr::collect()} from \code{dbplyr} after a simple
filter statements to speed up computations. Valid options are
\code{"subgroups"}, \code{"filters"}, or \code{"preprocess"}. Note that
\code{dbplyr} does not support all expressions.}
}
\value{
a nested \code{data.frame} containing all combinations of arbitrary
decisions for a multiverse analysis. Decision types will become list
columns matching the type of decisions called along the pipeline (e.g.,
filters, variables, etc.). Any decisions containing
\code{\link[glue]{glue}} syntax will be populated with the relevant
information.
}
\description{
Expand a set of multiverse decisions into all possible combinations
}
\examples{

library(tidyverse)
library(multitool)

the_data <-
  data.frame(
    id   = 1:500,
    iv1  = rnorm(500),
    iv2  = rnorm(500),
    iv3  = rnorm(500),
    mod1 = rnorm(500),
    mod2 = rnorm(500),
    mod3 = rnorm(500),
    cov1 = rnorm(500),
    cov2 = rnorm(500),
    dv1  = rnorm(500),
    dv2  = rnorm(500),
    include1 = rbinom(500, size = 1, prob = .1),
    include2 = sample(1:3, size = 500, replace = TRUE),
    include3 = rnorm(500)
  )

full_pipeline <-
  the_data |>
  add_filters(include1 == 0,include2 != 3,include2 != 2, include3 > -2.5) |>
  add_variables("ivs", iv1, iv2, iv3) |>
  add_variables("dvs", dv1, dv2) |>
  add_variables("mods", starts_with("mod")) |>
  add_preprocess(process_name = "scale_iv", 'mutate({ivs} = scale({ivs}))') |>
  add_preprocess(process_name = "scale_mod", mutate({mods} := scale({mods}))) |>
  add_summary_stats("iv_stats", starts_with("iv"), c("mean", "sd")) |>
  add_summary_stats("dv_stats", starts_with("dv"), c("skewness", "kurtosis")) |>
  add_correlations("predictors", matches("iv|mod|cov"), focus_set = c(cov1,cov2)) |>
  add_correlations("outcomes", matches("dv|mod"), focus_set = matches("dv")) |>
  add_reliabilities("unp_scale", c(iv1,iv2,iv3)) |>
  add_model("no covariates", lm({dvs} ~ {ivs} * {mods})) |>
  add_model("with covariates", lm({dvs} ~ {ivs} * {mods} + cov1)) |>
  add_postprocess("aov", aov())

pipeline_expanded <- expand_decisions(full_pipeline)
}
