% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitool-package.R
\docType{package}
\name{multitool-package}
\alias{multitool}
\alias{multitool-package}
\title{multitool: Tools for Running Multiverse Style Analyses}
\description{
Multiverse analyses run the same analysis over a range of arbitrary data processing decisions. 'multitool' provides an interface for creating alternative analysis pipelines and turning then into a grid of all possible pipelines. Using this grid as a blueprint, you can model your data with all possible analysis pipelines and summarize the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ethan-young.github.io/multitool/}
}

}
\author{
\strong{Maintainer}: Ethan Young \email{young.ethan.scott@gmail.com} (\href{https://orcid.org/0000-0002-8232-0184}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Stefan Vermeent \email{p.c.s.vermeent@gmail.com} (\href{https://orcid.org/0000-0002-9595-5373}{ORCID})
}

}
\keyword{internal}
