% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-grids.R
\name{post_filter_code}
\alias{post_filter_code}
\title{Add arbitrary code to execute after data are filtered (but before analysis)}
\usage{
post_filter_code(...)
}
\arguments{
\item{...}{the literal code you would like to execute after data are
filtered. \code{glue} syntax is allowed. An example might be centering or
scaling a predictor after the appropriate filters are applied to the data.}
}
\value{
a \code{tibble} with  two columns, "step" and "code". Each row
representing a post filtering step to execute. The "step" column indicates
the order and the "code" column is the literal code to be ran.
}
\description{
Add arbitrary code to execute after data are filtered (but before analysis)
}
\examples{
library(multitool)

post_filter_code(mutate({iv} := scale({iv}) |> as.numeric()))

}
