% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-multiverse.R
\name{report_universe_console}
\alias{report_universe_console}
\title{Create a 'universe' console report}
\usage{
report_universe_console(multiverse, my_data, decision_num)
}
\arguments{
\item{multiverse}{a \code{tibble} created from \code{\link{run_multiverse}}}

\item{my_data}{the original data used in the multiverse dataset}

\item{decision_num}{which decision set to create a report for}
}
\description{
Create a 'universe' console report
}
\examples{

\dontrun{
my_multi_results <- run_multiverse(my_grid, my_data)

report_universe_console(my_multi_verse, my_data, 1)
}
}
