% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-multiverse.R
\name{run_descriptives}
\alias{run_descriptives}
\title{Run a multiverse-style descriptive analysis based on a complete decision grid}
\usage{
run_descriptives(.pipeline, show_progress = FALSE)
}
\arguments{
\item{.pipeline}{a \code{tibble} produced by a series of \code{add_*} calls.
Importantly, this needs to be a pre-expanded pipeline because descriptive
analyses only change when the underlying cases change. Thus, only filtering
decisions will be used and internally expanded before calculating various
descriptive analyses.}

\item{show_progress}{logical, whether to show a progress bar while running.}
}
\value{
single \code{tibble} containing tidied results for all descriptive
analyses specified
}
\description{
Run a multiverse-style descriptive analysis based on a complete decision grid
}
\examples{
\dontrun{
run_descriptives(pipeline)
}
}
