% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-multiverse.R
\name{run_multiverse}
\alias{run_multiverse}
\title{Run a multiverse based on a complete decision grid}
\usage{
run_multiverse(my_grid, my_data, save_model = FALSE)
}
\arguments{
\item{my_data}{a \code{data.frame} representing the the original data}

\item{grid}{a \code{tibble} produced by \code{\link{combine_all_grids}}}
}
\value{
various thing based on the grid, model, and post hoc analyses
}
\description{
Run a multiverse based on a complete decision grid
}
\examples{

\dontrun{
run_multiverse(data, grid)
}

}
