% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-multiverse.R
\name{run_universe}
\alias{run_universe}
\title{Run a single set of arbitrary decisions and save the result}
\usage{
run_universe(my_grid, my_data, decision_num, save_model = FALSE)
}
\arguments{
\item{my_data}{a \code{data.frame} representing the the original data}

\item{decision_num}{an single integer from 1 to \code{nrow(grid)} indicating
which specific decision set to run}

\item{grid}{a \code{tibble} produced by \code{\link{combine_all_grids}}}
}
\value{
a single row \code{tibble} containing the decision, code that was
ran, the results, and any notes (e.g., warnings or messages).
}
\description{
Run a single set of arbitrary decisions and save the result
}
\examples{

\dontrun{
run_universe(my_grid, my_data, decision_num)
}
}
