% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-pipelines.R
\name{show_pipe_filter}
\alias{show_pipe_filter}
\alias{show_pipe_preprocess}
\alias{show_pipe_model}
\alias{show_pipe_postprocess}
\title{Show multiverse data code pipelines}
\usage{
show_pipe_filter(multiverse, decision_num, copy = F)

show_pipe_preprocess(multiverse, decision_num, copy = F)

show_pipe_model(multiverse, decision_num, mod, copy = F)

show_pipe_postprocess(multiverse, decision_num, mod, copy = F)
}
\arguments{
\item{multiverse}{a \code{tibble} created by \code{\link{run_multiverse}}}

\item{decision_num}{numeric. Indicates which 'universe' in the multiverse to
show underlying code.}

\item{copy}{logical. Whether to copy the pipeline code to the clipboard using
\code{\link[clipr]{write_clip}}.}

\item{mod}{the name of the model (e.g., \code{lm}) or post-processing task}
}
\description{
Each \code{show_pipeline*} function should be self-explanatory - they indicate
where along the multiverse pipeline to extract code. The goal of these
functions is to create a window into each multiverse decision set
context/results and allow the user to inspect specific decisions straight from
the code that produced it.
}
\section{Functions}{
\itemize{
\item \code{show_pipe_preprocess}: Show the code up to the preprocessing stage
'@export

\item \code{show_pipe_model}: Show the code up to the modeling stage

\item \code{show_pipe_postprocess}: Show the code up to the post_processing
stage
}}

